# api/swagger.yaml
basePath: /web
definitions:
...
  web.DuplicateCheckRequest:
    properties:
      amount:
        example: 12.5
        type: number
      date:
        example: "2026-09-03"
        type: string
      description:
        example: Irish pub with Laura
        type: string
    type: object
  web.DuplicateCheckResponse:
    properties:
      duplicates:
        items:
          $ref: '#/definitions/web.TransactionDTO'
        type: array
    type: object
...
  web.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  web.ParseTransactionRequest:
    properties:
      text:
        example: irish pub with laura 12.50 yesterday
        type: string
    type: object
  web.ParseTransactionResponse:
    properties:
      amount:
        example: 12.5
        type: number
      category:
        example: EatingOut
        type: string
      date:
        example: "2026-09-03"
        type: string
      description:
        example: Irish pub with Laura
        type: string
      duplicates:
        items:
          $ref: '#/definitions/web.TransactionDTO'
        type: array
      type:
        example: Expense
        type: string
    type: object
...
  web.TransactionDTO:
    properties:
      amount:
        type: number
      category:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: integer
      type:
        type: string
    type: object
...
info:
  contact: {}
  description: 'Personal finance tracking API. All /api/* endpoints require either
    a session cookie (browser) or an Authorization: Bearer <token> header (machine
    clients).'
  title: Pago API
  version: "1.0"
paths:
...
  /api/transactions/duplicates:
    post:
      consumes:
      - application/json
      description: Returns recent transactions (last 60 days before the given date)
        with the same amount and a matching description.
      parameters:
      - description: Candidate transaction
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.DuplicateCheckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.DuplicateCheckResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Find possible duplicates
      tags:
      - transactions
...
  /api/transactions/parse:
    post:
      consumes:
      - application/json
      description: Uses the configured LLM to extract type, amount, category, description
        and date from a short sentence. Nothing is saved. Also returns recent transactions
        that look like duplicates.
      parameters:
      - description: Free text
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.ParseTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.ParseTransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Parse free text into a transaction (AI)
      tags:
      - transactions
...
securityDefinitions:
  BearerAuth:
    description: 'Provide the value as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"